move genrateNNodes out of test file
